moving logging related stuff to it's own file
